refactor: DatabaseInfo.java to LoginBuilder.java and moved it into a builder package
